Create rocknsm.spec
all magic included at the top. I couldn't find out how to awk the variables inside of the spec file so there is a few curls that ultimately are unnecessary if the variable could be parsed to get the next link instead. 

Errors from rpmlint all disliked the placement of the files in /opt. No other errors were detected.