Create es-parse-failures-mappings.json
Mappings for the parse failure index from the other logstash pull requests.
This will enablemapping all fields as a "keyword" so as not to use any elasticsearch default "coerce"(ing) and "standard" analyzers.
The only type specified is "@timestamp" which is a date ---- of when the log was received.
Also, set to not replicate ("number_of_replicas": 0) as no need to take up space for parse failures/errors.
Note when testing "POST"ing / adding to this index that it won't refresh in Kibana/Elasticsearch until 45 seconds (specified in "refresh_interval")