Fixed #37 by using  include_vars for rocknsm config
The playbooks/sensors.yml and playbooks/servers.yml were both missing the following line:

include_vars: /etc/rocknsm/config.yml

Simply adding this as a task didn’t fix the problem because the order of execution when using roles. After reading http://docs.ansible.com/ansible/latest/playbooks_reuse_roles.html it was clear that pre_tasks must be used in order to be processed before roles.

For referference, the include_vars config was in the playbooks/all.yml file, but was not in the other files.

To fix the problem, added the pre_tasks to load the variables so they would be loaded when processing the playbook