Merge remote-tracking branch 'tfplenum/devel' into tfplenum-devel
Set Kibana reverse proxy config. Moved portainer to port 8443 on proxy as it does not have the ability to work with rewritten paths. Added Certificate Authority generation on haproxy with signed server cert for ssl termination. Client certs for sensors identified in the inventory file are created for future implementation.