Fix to comply with @bastionbot's commit (#81)
* Updated inventory file

* Merge branch 'haproxy' into devel

Changed haproxy conf to rewrite urls

* Merge remote-tracking branch 'tfplenum/devel' into tfplenum-devel

Set Kibana reverse proxy config. Moved portainer to port 8443 on proxy as it does not have the ability to work with rewritten paths. Added Certificate Authority generation on haproxy with signed server cert for ssl termination. Client certs for sensors identified in the inventory file are created for future implementation.

* openssl.cnf template got lost in the last haproxy commit.

* Implemented prompts to select monitoring interfaces via ansible pause task.

* Elasticsearch wait task

Corrected Elasticsearch wait task to work with local/multinode and haproxy. However it is dependent on at least one server having the haproxy role

* Fix to comply with @bastionbot's commit