Set template order to higher priority
This sets the stage for adding a default template with lower priority.

Add default template

Set order to 20 to allow for default and ECS templates

Add latest mapping for ECS

This is mostly a placeholder, but is a first step on the path to full
adoption.

Add tasks for default and ECS index templates

Add missing comma

Change doc type to `_doc` vs `doc`

Update ES templates