Update logrotate configuration
Thanks to @leanvel for the great research in Issue #170. I set the selinux type of the various data dirs to `var_log_t`. There's a number of options available, but after reading the docs that made the most sense.

Also, I found that `logrotate` is supposed to already have permission to send a `reload` to any systemd service unit. Our suricata service unit already maps this to the SIGHUP signal, so I swapped out this config to work within the existing confines of the selinux policy.

**NOTE**: Still need to test this, but I think it covers all the bases.

Fixes #170

Co-Authored-By: LeanVel <leanvel@users.noreply.github.com>